+++ src/types.ts
@@ -1,10 +1,12 @@
 import { GraphQLError } from "graphql"
+import { FetchMoreOptions, FetchMoreQueryOptions } from "apollo-client"
 
 export interface QueryAttributes<Data = {}, Variables = {}> {
   data: Data | null | undefined
   errors: GraphQLError[] | null | undefined
   variables: Variables
   loading: boolean
+  fetchMore: (options: FetchMoreOptions & FetchMoreQueryOptions) => void
   refetch: () => void
 }